Working on convolutions

+++ src/series/chunksize.h
@@ -4,4 +4,4 @@
 
 #include "defs/CHUNK_SIZE_LOG2.h"
 static_assert(CHUNK_SIZE_LOG2 < sizeof(unsigned int) * CHAR_BIT, "Lots of things depend on an inter-chunk index fitting into an unsigned int");
-#define CHUNK_SIZE (static_cast<unsigned int>(1) << CHUNK_SIZE_LOG2)
+#define CHUNK_SIZE (1u << CHUNK_SIZE_LOG2)
